+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **\[Python/C++\]** Add `CheckAndUpdateBounds` to `LasHeader` to check values in a `Point` and update the bounds if necessary.
 - **\[Python/C++\]** Add `SetGpsTimeBit` to `LasHeader` to easily set the bit value to 1.
 
+### Fixed
+- **\[Python/C++\]** Change return type of `LasHeader::ApplyInverseScale` from `double` to `uint64_t`.
+
 ## [2.2.3] - 2022-02-02
 
 ### Fixed